% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parler_auth.R
\name{parler_auth}
\alias{parler_auth}
\title{Authenticate}
\usage{
parler_auth(
  decode = TRUE,
  path = "config/auth.json",
  option = "enter",
  mst = NULL,
  jst = NULL
)
}
\arguments{
\item{decode}{Whether MST and JST tokens should be decoded before storing. \code{BOOLEAN} defaulting to \code{TRUE}.}

\item{path}{Path where to store the authentication file. Currently needs to stay './config/auth.json'.}

\item{option}{How JST and MST tokens shall be entered. \code{str} defaulting to 'enter'.}
}
\description{
xxx
}
\details{
Warning: This functions writes an authentication file to a json-file in
./confic/auth.json
}
\examples{
\dontrun{
# parler_auth()
}
}
