% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parler_hashtag.R
\name{parler_hashtag}
\alias{parler_hashtag}
\title{Get Parler Posts by Hashtag}
\usage{
parler_hashtag(
  hashtag,
  output_format = "data.frame",
  flatten_sep = " || ",
  parse_numbers = TRUE,
  verbose = FALSE,
  timeout = 600
)
}
\arguments{
\item{hashtag}{Parler user handle.}

\item{output_format}{Options include: \code{data.frame} (flattened, one row per post), \code{list}, and \code{json} (raw json file)}

\item{flatten_sep}{Separator used to flatten nested variables (e.g. multiple badges per user). Defaults to \code{||}}

\item{parse_numbers}{Whether to parse numbers from Parler's format to numeric (e.g. from \verb{16k} to 16000)}

\item{verbose}{Whether to print additional information while scraping}

\item{timeout}{Whether to stop the scraping after an amount of timeout seconds.}
}
\value{
A dataframe with one row for each post and a column for each variable.
}
\description{
Gets Parler Posts by hashtag
}
\details{
Warning: ...
}
\examples{
\dontrun{
out <-
  parler_hashtag(hashtag  = "nocovidvaccine",
                 output_format = "data.frame",
                 flatten_sep = " || ",
                 parse_numbers = TRUE,
                 verbose = TRUE)

print(colnames(out))
}
}
