% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parler_posts.R
\name{parler_posts}
\alias{parler_posts}
\title{Get Parler User Posts}
\usage{
parler_posts(
  user,
  output_format = "data.frame",
  flatten_sep = " || ",
  parse_numbers = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{user}{Parler user handle.}

\item{output_format}{Options include: \code{data.frame} (flattened, one row per post), \code{list}, and \code{json} (raw json file)}

\item{flatten_sep}{Separator used to flatten nested variables (e.g. multiple badges per user). Defaults to \code{||}}

\item{parse_numbers}{Whether to parse numbers from Parler's format to numeric (e.g. from \verb{16k} to 16000)}

\item{verbose}{Whether to print additional information while scraping}
}
\value{
A dataframe with one row for each post and a column for each variable.
}
\description{
Gets Posts by username
}
\details{
Warning: ...
}
\examples{
\dontrun{
out <-
  parler_posts(user = "Grenell",
                 output_format = "data.frame",
                 flatten_sep = " || ",
                 parse_numbers = TRUE,
                 verbose = TRUE)

print(colnames(out))
}
}
