% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parler_profile.R
\name{parler_profile}
\alias{parler_profile}
\title{Get Parler Profile Data}
\usage{
parler_profile(
  user,
  output_format = "data.frame",
  flatten_sep = " || ",
  token_variables = FALSE,
  parse_numbers = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{user}{Parler user handle.}

\item{output_format}{Options include: \code{data.frame} (flattened), \code{list}, and \code{json} (raw json file)}

\item{flatten_sep}{Separator used to flatten nested variables (e.g. multiple badges per user). Defaults to \code{||}}

\item{token_variables}{Whether to output information associated with token (e.g. follower relationship between token user and scraped user). Defaults to \code{FALSE}}

\item{parse_numbers}{Whether to parse numbers from Parler's format to numeric (e.g. from \verb{16k} to 16000)}

\item{verbose}{Whether to print additional information while scraping}
}
\value{
A dataframe with one row and a column for each variable.
}
\description{
Gets Relevant Profile/User Metadata by username
}
\details{
Warning: ...
}
\examples{
\dontrun{
out <-
  parler_profile(user = "caseybmulligan",
                 output_format = "data.frame",
                 flatten_sep = " || ",
                 token_variables = FALSE,
                 parse_numbers = TRUE,
                 verbose = TRUE)

print(colnames(out))
}
}
