% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_export.R
\name{parler_time}
\alias{parler_time}
\title{Parse Parler Time Format}
\usage{
parler_time(parltime, out_format = "character")
}
\arguments{
\item{parltime}{A parler timestamp}

\item{out_format}{Either \code{character} (ISO8601) or \code{POSIXlt}}
}
\description{
Parse Time Format from Parler
}
\details{
...
}
\examples{
parler_time(parltime = "20201015232824",
            out_format = "character")
}
